Refactor command structure and improve project organization
- Replaced UI command references with Commands namespace in ListWithNoData, PluginDataBaseGameBase, and PluginSettingsViewModel for better modularity.
- Updated project file to reflect new command file locations, enhancing project structure.
- Removed obsolete UI command files (CommandsNavigation, CommandsSettings, CommandsWindows) to streamline the codebase and reduce redundancy.
- Added new cache management methods in LiteDbItemCollection for improved session handling.
